Reformat code in functions, made first nodes of the decisional tree, reformat code to work on nodes of type struct
best_move.cpp contains a modified version of move.cpp, in which I managed to reformat the code to be all contained in functions and then added other functions to help me create the decisional tree. Mainly added evaluation and make tree function, the latter only adding to the tree the first set of big children branching from the given state. best_move2.cpp is the improved version of best_move.cpp which uses a struct for the nodes.